Move submit & reset button out of ConfigForm; Rename ConfigForm to ConfigDialog
